src/main/java/io/sampathsl/oauth2/demo/jwtdemo/controller/UserController.java
package io.sampathsl.oauth2.demo.jwtdemo.controller;

import java.util.List;

import io.sampathsl.oauth2.demo.jwtdemo.model.User;
import io.sampathsl.oauth2.demo.jwtdemo.service.UserService;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/v1")
@RequiredArgsConstructor
public class UserController {

  private final UserService userService;

  @GetMapping("/users")
  public ResponseEntity<List<User>> getAllUsers() {
    List<User> users = userService.getAllUsers();
    return new ResponseEntity<>(users, HttpStatus.OK);
  }
}
